% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateScenario.R
\name{evaluateScenario}
\alias{evaluateScenario}
\title{Evaluate a subsetting scenario}
\usage{
evaluateScenario(x, subset, k)
}
\arguments{
\item{x}{Numeric matrix containing the dataset of interest.
Rows are dimensions while columns are observations.}

\item{subset}{Vector specifying the subset of observations in \code{x} to use as the initial dataset.}

\item{k}{Integer scalar specifying the number of clusters.}
}
\value{
Data frame containing the WCSS and Rand indices for each strategy.
}
\description{
Evaluate a subsetting scenario by computing WCSS and Rand indices to compare clusterings.
}
\details{
The idea is to obtain an initial clustering on \code{x[,subset]} and then apply various strategies to obtain a clustering on \code{x}.
This can be:
\itemize{
\item \code{rest}, where the centers from the initial clustering are used directly for Hartigan-Wong iterations with the full dataset.
\item \code{full}, where a fresh initialization is performed with all observations (not using any information from the initial clustering).
Hartigan-Wong iterations are then performed as usual.
\item \code{reinit}, where a reinitialization strategy is applied that attempts to preserve information from the existing clustering. 
Hartigan-Wong iterations are then performed as usual.
}
We compute the WCSS of the clusterings obtained with each strategy.
We also compute the Rand index of each clustering with the initial clustering of the subset.
}
\examples{
populations <- matrix(rnorm(100, sd=5), ncol=10)
sampled <- sample(ncol(populations), 1000, replace=TRUE)
stuff <- populations[,sampled]
stuff <- stuff + rnorm(length(stuff))
                                                         
sub <- sampled < ncol(populations)
evaluateScenario(stuff, sub, 10)

}
