% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randIndex.R
\name{randIndex}
\alias{randIndex}
\title{Compute the adjusted rand index}
\usage{
randIndex(x, y)
}
\arguments{
\item{x}{Vector or factor containing the first clustering.}

\item{y}{Vector or factor containing the second clustering, of the same length as \code{x}.}
}
\value{
Numeric scalar containing the adjusted Rand index.
}
\description{
Compute the adjusted Rand index to quantify the similarity between two clusterings.
}
\examples{
mat <- t(as.matrix(iris[,1:4]))

first <- runKmeans(mat, centers=5)
second <- runKmeans(mat, centers=5)
randIndex(first$clusters, second$clusters)

third <- runKmeans(mat, centers=10)
randIndex(first$clusters, third$clusters)

}
