% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKmeans.R
\name{runKmeans}
\alias{runKmeans}
\title{Run k-means with existing centers}
\usage{
runKmeans(x, centers, reinitialize = FALSE)
}
\arguments{
\item{x}{Numeric matrix of dimensions (rows) and observations (columns).}

\item{centers}{Numeric matrix of dimensions (rows) and centers (columns).
Alternatively a numeric scalar specifying the number of centers to use.}

\item{reinitialize}{Logical scalar indicating whether reinitialization should be used when \code{centers} is a matrix.
Otherwise, the provided centers are used directly for further iterations.}
}
\value{
A list containing the result of k-means clustering on observations of \code{x}.
}
\description{
Run k-means with various strategies for handling existing centers.
}
\examples{
mat <- t(as.matrix(iris[,1:4]))
sub <- mat[,1:100]

# Initial run:
ref <- runKmeans(sub, centers=5)

# Using the existing clusters for the full dataset.
rest <- runKmeans(mat, centers=ref$centers)

# Using the reinitialization strategy instead.
reinit <- runKmeans(mat, centers=ref$centers, reinitialize=TRUE)

}
\author{
Aaron Lun
}
