% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateScenario.R
\name{simulateScenario}
\alias{simulateScenario}
\title{Simulate scenarios and evaluate them}
\usage{
simulateScenario(FUN, k, iterations = 100)
}
\arguments{
\item{FUN}{Function that accepts no arguments and returns a list with 
\code{data}, a numeric data matrix with dimensions and observations in rows and columns, respectively;
and \code{subset}, a logical vector specifying the subset to retain in the initial clustering.}

\item{k}{Integer scalar specifying the number of clusters to return.}

\item{iterations}{Integer scalar specifying the number of iterations to perform.}
}
\value{
A matrix of WCSS and Rand indices for each strategy (columns) in each iteration (row).
}
\description{
Generate simulated data and collate statistics for them.
}
\examples{
FUN <- function() {
    populations <- matrix(rnorm(100, sd=5), ncol=10)
    sampled <- sample(ncol(populations), 1000, replace=TRUE)
    stuff <- populations[,sampled]
    stuff <- stuff + rnorm(length(stuff))
    list(data=stuff, subset=(sampled < ncol(populations)))
}

out <- simulateScenario(FUN, 10)
plot(out$wcss, out$rand, col=factor(out$method))

}
\author{
Aaron Lun
}
