% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeResults.R
\name{summarizeResults}
\alias{summarizeResults}
\alias{plotResults}
\title{Summarize simulation results}
\usage{
summarizeResults(df, collapse = FALSE)

plotResults(df)
}
\arguments{
\item{df}{A data frame of WCSS and Rand statistics, usually generated from \code{\link{simulateScenarios}}.}

\item{collapse}{Logical scalar indicating whether the results should be collapsed ito a single data frame.}
}
\value{
For \code{summarizeResults}, a list of data frames containing the mean and standard error of the Rand index and WCSS for each method.
If \code{collapse=TRUE}, the statistics are rearranged into a single data.frame with a single row. 

For \code{plotResults}, a list of ggplot objects containing scatter and violin plots.
}
\description{
Summarize and plot the simulation results.
}
\examples{
FUN <- function() {
    populations <- matrix(rnorm(100, sd=5), ncol=10)
    sampled <- sample(ncol(populations), 1000, replace=TRUE)
    stuff <- populations[,sampled]
    stuff <- stuff + rnorm(length(stuff))
    list(data=stuff, subset=(sampled < ncol(populations)))
}

out <- simulateScenario(FUN, 10)
summarizeResults(out)

plots <- plotResults(out)
plots$both
plots$rand
plots$wcss

}
\author{
Aaron Lun
}
